feat: remove msg buffer and socket client offset
This change affects the handling of message events,
reverting it to default handling. Delivery under
bad connection becomes less reliable, but code is
greatly simplified, and a serious problem with the
code is avoided.

The removed code was previously written to make
transmission of message events to server and
callbacks from server more reliable under bad
connection.

But, the removed code was very difficult to
implement and possesses problems i could not fix.

Problems observed:
1. Server keeps the same socket client offset for all
sockets.
2. Reload of client resets the sent message id ref.

Furthermore, it greatly complicates the socket io
code and made further developments difficult.

Thus, it was removed in favor of the default socket
io guarantee of delivery solution. The default relies
on sendBuffer, and is not as reliable but it is simple.

This will be the path from now onwards, until i figure
out how to reimplement the removed code.