feat: add validation for message event and refactor
Validation for message event is done by checking
if user is registered for the group it is posting
to.

Connection handler is refactored to store user data
in socket for later reference.
Wait on initialization is done before registration
of listeners are allowed.

Various functions and code sections in comm() is
moved out as external functions.