feat: change direct chat url and database format
Direct chat url now uses a unique id for direct chats
instead of using friend user_id.
This enables consistent id for both users of the
direct chat, simplifying the code and also allowing
users to share their url and get the same results.

To achieve this, database is refactored to move the
relevant data to a separate direct_chat table and
it's accompanying agent table.

Heavy refactor is done as a consequence of changing
database format.
Column names for foreign key references is corrected
on plurality rules.
Private chat is now refered to as direct chat.