fix: restrict access and gracefully handle exceptions
1. Restrict access to url api.

Group and direct chat info and messages will only
be provided if the fetching user has permissions.
Else, false boolean is returned to the fetch request.

2. Gracefully handle failure to deserialize user

if fetches for user info fails, return false.
if there is no cookies, avoid try to call the fetch
as the handle isn't defined, return false.

On failure to retrieve user info in socket io,
disconnect the socket and quit the handler.