refactor: define a class for chat item data
Last message of chat item data is now defined with
a default instead of undefined.
Changes are made to identify undefined last message
by checking id of last message to see if it is
default.

Move sortChats to ChatItemData as a static method.

Enable inputs into ChatItemData to be automatically
cast to the relevant class.