feat: add direct chat buttons to friend overview
Show chat button is shown if direct chat of user
is set to not shown or not created.
It will created direct chat or set it to shown.

A link to direct chat is shown instead if direct
chat is set to shown.

When direct chat is shown, add new entry will be
added to chatlist.

fix spelling: sendRequest -> sentRequest