refactor: change name for queries parameter
For queries with parameter of 2 user ids, the change
is as follow:
sender_id -> user_id
receiver_id -> other_id

This change is to make the origin of the ids clearer.
user_id is the id of the requesting client.
other_id is the id of another user that the
requesting client is referring to.

The comment in Room.jsx for handleUnfriend() is
also modified.