fix: remove nesting of chat button within link
Chat button in friendItemButtonBar is nested within
a link. This is because it is a link but styled
by using a button element.

This isn't valid html, thus the nesting is removed,
and the link is styled with it's own css to look
like a button.

Styling for button-link class and other items is
modified.