fix: display contents correctly while loading
While loading, default values are used, showing confusing contents. Once
loading is complete, the change in content is visually annoying.

To prevent this, default values are changed to be non visible, so that
loading would fill in the blank spaces.

An empty array is used as default value for parameter accepting array values
instead of an array with a single instance.

E.G., parameter = [new Class()] --> parameter = []

This is to prevent items from appearing in UI when the array is empty.
The datatype info indicated by an instance in an array is replaced with
comments.

DEFAULT_NAME is changed to DEFAULT_TEXT and the value is set to zero
width space. The purpose of the constant is now to prevent text
container collapse while remaining hidden.