refactor(queries): improve readability
Remove getFriendsByUserId(), replacing calls to it with getFriendships()
to simplify queries. Modify getFriendships() to allow query based on
friendship state.

Change how node-postgres query arguments(text, values) is defined in
functions where it is dependent on parameters, to make it clearer.

Add various comments on the functions and implementations in queries.js