style: change layout of chatroom for narrow viewport
The width breakpoint for layout change is set to minimum width where the
chatroom can be displayed reasonably well.

Hide low priority display class elements in chatroom when breakpoint is
reached(either chatlist or roomui).

narrow-width and wide-width classes are used as selectors for setting
width in chatroom.

Roomui is given the low priority class when room info is displayed,
while chatlist is always low priority.

ChatItemMenu is still displayed when chatlist is hidden, it should be
hidden as well. Future commit to change ChatItemMenu will resolve this.