fix: solve ChatItemMenu responsiveness issues
The position of the menu fails to update upon reorder of chatlist items
and on scroll.

Use interval to update ChatItem rect values, this makes the menu
responsive to scroll and position changes.

Change positioning of menu to absolute and contained on local block.
This anchors the menu to ChatItem, making it's positioning less
problematic.

Move the menu tag to the location after the menu button, this fixes
keyboard tab navigation which is previously disordered.

Close menu when open menu button is scrolled out of view. Stop changing
menu's horizontal position, permanently set it to left of button.

Change positioning of menu from top/left to bottom/right, to ensure the
default position of the menu does not produce scrollbars from auto
overflow.

Do some optimisation using useCallback, and don't update rect state if
rect properties are unchanged.

Stop blocking of ChatList scroll when menu is open.