feat: change the structure of the website
The index path is set to intro page to introduce the web application.
Home page is set to '/home' path instead.

Rename 'Layout' component to 'PrivateInterface' as it is an interface
for logged in users. Create an equivalent public interface for public
users.

Public links goes to public interface, while private links goes to
private interface. This is to separate related links into what seems
like 2 different websites. A website for logged in users with chat
functions, and another for public users with information and
login/signup links.

Allow 'Nav' to show different links depending the interface it is
rendered in. The links are different as logged in users will require and
have access to different functionalities.