feat: redirect to target url after login
When navigating to a target url that requires login, user will be
redirected to the login page. After login, user is redirected to the
home page instead of the target url.

This commit keeps the target url in the query string of the login url
upon redirect to login. Then uses the query string to redirect to target
url on successful authentication.