feat: show loading status
Show loading state for chat room and room info.

When "unfriend" event happens in chat room, redirect to home instead of
clearing data.

Create "LoadError" component to display error message. Used when fetch
is successful but returns an invalid result for various reasons.
Identified by fetch return of false when other values are expected.
State is stored with the false value to indicate this condition.