feat: add guest login
Guest login allows easy creation of a temporary account with only a
username.

It is identified by "guest_" in front of username. It is automatically
generated with 4 digits, but can also be entered manually.

This account uses session cookie instead of password for login, and will
be lost on clearing of cookies or switching to another device.