feat: store room state in URL query
Store in URL query whether room info is shown and whether it is
expanded. This preserves some room state to return to on reload,
navigation or link sharing.

Use display none instead of removing toggled hidden component in room.
It is desirable to preserve scroll position and other states on the
components by not removing it.

style:

 - expand room info on query change

refactor:

 - Create useSearchState hook to access react router's useSearchParams
 with a useState like interface. This makes for a more familiar
 interface.

 - move state clearing code on room change to useFetchedState hook.

 - create boolTypes to specify setting bool states to true, false or
 toggle.