fix: enable Socket.IO listeners to work properly
The Socket.IO bug fix from a previous commit did not work. The bug came
from running io() multiple times and overwriting previously returned
socket object.

Old listeners are assigned to old socket object, and is unreachable for
removal when socket is overwritten. This caused the listeners to appear
to disappear, but it is still running in the background with stale
functions.

io() is now called once, with the returned socket used to establish
connection in the loader where io() is originally declared.