feat: sync user's connected clients
When a user opens the website in different tabs or devices, some of
their states need to be synced. Socket.IO emit is directed to all
connected sockets sharing the same user_id.

User status is kept online until all of the user's clients are
disconnected.

Refactor is done, changing some api to sockets.